moved tests folder all working
