+++ .npmignore
@@ -1,7 +1,7 @@
 # Source directory
 src/
 # Tests directory
-#tests/
+tests/
 # Coverage directory
 coverage/
 # Code Documentation directory
