+++ .vscode/launch.json
@@ -15,7 +15,7 @@
                 "--timeout",
                 "999999",
                 "--colors",
-                "${workspaceFolder}/src/tests/json_objects.test.ts",
+                "${workspaceFolder}/tests/json_objects.test.ts",
             ],
             "internalConsoleOptions": "neverOpen",
             "protocol": "inspector",
@@ -31,7 +31,7 @@
                 "--timeout",
                 "999999",
                 "--colors",
-                "${workspaceFolder}/src/tests/pools.test.ts",
+                "${workspaceFolder}/tests/pools.test.ts",
             ],
             "internalConsoleOptions": "neverOpen",
             "protocol": "inspector",
@@ -47,7 +47,7 @@
                 "--timeout",
                 "999999",
                 "--colors",
-                "${workspaceFolder}/src/tests/outside_files.tests.ts",
+                "${workspaceFolder}/tests/outside_files.tests.ts",
             ],
             "internalConsoleOptions": "neverOpen",
             "protocol": "inspector",
@@ -63,7 +63,7 @@
                 "--timeout",
                 "999999",
                 "--colors",
-                "${workspaceFolder}/src/tests/devCloud_config.tests.ts",
+                "${workspaceFolder}/tests/devCloud_config.tests.ts",
             ],
             "internalConsoleOptions": "neverOpen",
             "protocol": "inspector",
@@ -79,7 +79,7 @@
                 "--timeout",
                 "999999",
                 "--colors",
-                "${workspaceFolder}/src/tests/devCloud_ucs.tests.ts",
+                "${workspaceFolder}/tests/devCloud_ucs.tests.ts",
             ],
             "internalConsoleOptions": "neverOpen",
             "protocol": "inspector",
@@ -95,7 +95,7 @@
                 "--timeout",
                 "999999",
                 "--colors",
-                "${workspaceFolder}/src/tests/devCloud_qkview.tests.ts",
+                "${workspaceFolder}/tests/devCloud_qkview.tests.ts",
             ],
             "internalConsoleOptions": "neverOpen",
             "protocol": "inspector",
@@ -127,7 +127,7 @@
                 "--timeout",
                 "999999",
                 "--colors",
-                "${workspaceFolder}/src/tests/unPacker.tests.ts",
+                "${workspaceFolder}/tests/unPacker.tests.ts",
             ],
             "internalConsoleOptions": "neverOpen",
             "protocol": "inspector",
@@ -143,7 +143,7 @@
                 "--timeout",
                 "999999",
                 "--colors",
-                "${workspaceFolder}/src/tests/tmosParser.tests.ts",
+                "${workspaceFolder}/tests/tmosParser.tests.ts",
             ],
             "internalConsoleOptions": "neverOpen",
             "protocol": "inspector",
@@ -160,7 +160,7 @@
                 "--timeout",
                 "999999",
                 "--colors",
-                "${workspaceFolder}/src/tests/ALL.tests.ts",
+                "${workspaceFolder}/tests/ALL.tests.ts",
             ],
             "internalConsoleOptions": "neverOpen",
             "protocol": "inspector",
