+++ tests/devCloud_ucs.tests.ts
@@ -7,7 +7,7 @@ import assert from 'assert';
 import * as fs from 'fs';
 import * as path from 'path';
 
-import BigipConfig from '../ltm';
+import BigipConfig from '../src/ltm';
 import { logOutput } from './explosionOutput';
 
 /**
