+++ tsconfig.json
@@ -13,6 +13,7 @@
         "rootDir": "src",
         "outDir": "dist"
     },
+    "compileOnSave": true,
     "include": [
         "src/**/*"
     ],
